update new code

+++ flink-runtime/src/main/java/org/apache/flink/runtime/operators/resettable/SpillingResettableMutableObjectIterator.java
@@ -53,9 +53,9 @@ public class SpillingResettableMutableObjectIterator<T> implements ResettableMut
 	
 	protected final TypeSerializer<T> serializer;
 	
-	private int elementCount;
+	private long elementCount;
 	
-	private int currentElementNum;
+	private long currentElementNum;
 	
 	protected final SpillingBuffer buffer;
